requirements.txt
# Основные зависимости
requests>=2.31.0        # HTTP-запросы к XMLStock API
lxml>=4.9.0             # Парсинг XML от Яндекса
sqlalchemy>=2.0.0       # Работа с SQLite
pandas>=2.0.0           # Обработка данных для отчетов
jinja2>=3.1.0           # Шаблоны HTML отчетов
python-dotenv>=1.0.0    # Загрузка переменных окружения

# Для разработки (опционально)
pytest>=7.4.0           # Тестирование
black>=23.0.0           # Форматирование кода
flake8>=6.0.0           # Линтинг